+++ src/pages/assets/images-and-videos.js
@@ -15,7 +15,7 @@ import BulletedList from "../../components/BulletedList";
 export default () => (
   <Page>
     <Helmet>
-      <title>{`Images and videos`}</title>
+      <title>Images and videos</title>
     </Helmet>
     <Breadcrumbs
       items={[
@@ -30,27 +30,29 @@ export default () => (
       ]}
     />
     <Main>
-      <Title>{`Images and videos`}</Title>
-      <Paragraph>{`When creating images and videos you should:`}</Paragraph>
+      <Title>Images and videos</Title>
+      <Paragraph>When creating images and videos you should:</Paragraph>
       <BulletedList>
-        <ListItem>{`have consent from all subjects`}</ListItem>
-        <ListItem>{`have the subjects look into the camera`}</ListItem>
-        <ListItem>{`use natural light in a real location`}</ListItem>
-        <ListItem>{`avoid using actors`}</ListItem>
-        <ListItem>{`avoid adding soundtracks to videos`}</ListItem>
+        <ListItem>have consent from all subjects</ListItem>
+        <ListItem>have the subjects look into the camera</ListItem>
+        <ListItem>use natural light in a real location</ListItem>
+        <ListItem>avoid using actors</ListItem>
+        <ListItem>avoid adding soundtracks to videos</ListItem>
       </BulletedList>
-      <Paragraph>{`Images and videos should show:`}</Paragraph>
+      <Paragraph>Images and videos should show:</Paragraph>
       <BulletedList>
-        <ListItem>{`diversity`}</ListItem>
-        <ListItem>{`strength`}</ListItem>
-        <ListItem>{`resilience`}</ListItem>
-        <ListItem>{`positivity`}</ListItem>
-        <ListItem>{`genuineness`}</ListItem>
+        <ListItem>diversity</ListItem>
+        <ListItem>strength</ListItem>
+        <ListItem>resilience</ListItem>
+        <ListItem>positivity</ListItem>
+        <ListItem>genuineness</ListItem>
       </BulletedList>
       <Cite>
-        {`View `}
-        <Link href="https://github.com/barnardos/design-system/issues/227">{`images and videos research`}</Link>
-        {` on GitHub.`}
+        View{" "}
+        <Link href="https://github.com/barnardos/design-system/issues/227">
+          images and videos research
+        </Link>{" "}
+        on GitHub.
       </Cite>
     </Main>
     <Pagination