+++ src/pages/standards/typography.js
@@ -23,7 +23,7 @@ import TypeFigures from "../../components/TypeFigures";
 export default () => (
   <Page>
     <Helmet>
-      <title>{`Typography`}</title>
+      <title>Typography</title>
     </Helmet>
     <Breadcrumbs
       items={[
@@ -38,8 +38,8 @@ export default () => (
       ]}
     />
     <Main>
-      <Topic>{`Design`}</Topic>
-      <Title>{`Typography`}</Title>
+      <Topic>Design</Topic>
+      <Title>Typography</Title>
       <Contents>
         <ContentsMenu
           items={[
@@ -71,58 +71,65 @@ export default () => (
           ]}
         />
         <Content>
-          <Paragraph>{`You should use:`}</Paragraph>
+          <Paragraph>You should use:</Paragraph>
           <BulletedList>
-            <ListItem>{`a base size of 1rem (roughly 16px) for body copy`}</ListItem>
-            <ListItem>{`exponents of 1.125 for other type sizes (major second type scale)`}</ListItem>
+            <ListItem>
+              a base size of 1rem (roughly 16px) for body copy
+            </ListItem>
+            <ListItem>
+              exponents of 1.125 for other type sizes (major second type scale)
+            </ListItem>
           </BulletedList>
           <Paragraph>
-            {`Typography is subject to the `}
-            <Link href="/standards/viewport-scale/">{`viewport scale`}</Link>
-            {`.`}
+            Typography is subject to the{" "}
+            <Link href="/standards/viewport-scale/">viewport scale</Link>.
           </Paragraph>
-          <Heading>{`Type scale`}</Heading>
+          <Heading>Type scale</Heading>
           <Figures>
             <TypeFigures />
           </Figures>
           <Cite>
-            {`View `}
-            <Link href="https://github.com/barnardos/design-system/blob/master/src/custom-properties.css#L5-L22">{`type scale code`}</Link>
-            {` and `}
-            <Link href="https://github.com/barnardos/design-system/issues/3">{`type scale research`}</Link>
-            {` on GitHub.`}
+            View{" "}
+            <Link href="https://github.com/barnardos/design-system/blob/master/src/custom-properties.css#L5-L22">
+              type scale code
+            </Link>{" "}
+            and{" "}
+            <Link href="https://github.com/barnardos/design-system/issues/3">
+              type scale research
+            </Link>{" "}
+            on GitHub.
           </Cite>
-          <Heading>{`Line height`}</Heading>
-          <Paragraph>{`You should use multiples of 0.25rem (roughly 4px).`}</Paragraph>
-          <Heading>{`Line length`}</Heading>
-          <Paragraph>{`The optimal line length of text is around:`}</Paragraph>
-          <BulletedList>
-            <ListItem>{`15–40 characters for short lines`}</ListItem>
-            <ListItem>{`40–95 characters for body copy`}</ListItem>
-          </BulletedList>
-          <Heading>{`Brand font`}</Heading>
-          <Paragraph>{`You should use Proxima Nova.`}</Paragraph>
-          <Paragraph>{`You should only use the weights:`}</Paragraph>
-          <BulletedList>
-            <ListItem>{`semi-bold`}</ListItem>
-            <ListItem>{`light`}</ListItem>
-          </BulletedList>
+          <Heading>Line height</Heading>
           <Paragraph>
-            {`Use "`}
-            <Code>{`djy7snj`}</Code>
-            {`" for the `}
-            <Link href="https://helpx.adobe.com/uk/fonts/using/embed-codes.html">
-              {`Adobe Fonts embed code`}
-            </Link>
-            {`. This code is valid for all of Barnardo's products.`}
+            You should use multiples of 0.25rem (roughly 4px).
           </Paragraph>
-          <Heading>{`Body copy`}</Heading>
+          <Heading>Line length</Heading>
+          <Paragraph>The optimal line length of text is around:</Paragraph>
+          <BulletedList>
+            <ListItem>15–40 characters for short lines</ListItem>
+            <ListItem>40–95 characters for body copy</ListItem>
+          </BulletedList>
+          <Heading>Brand font</Heading>
+          <Paragraph>You should use Proxima Nova.</Paragraph>
+          <Paragraph>You should only use the weights:</Paragraph>
+          <BulletedList>
+            <ListItem>semi-bold</ListItem>
+            <ListItem>light</ListItem>
+          </BulletedList>
           <Paragraph>
-            {`You should use the `}
-            <Link href="https://www.w3.org/TR/css-fonts-4/#system-ui-def">
-              {`system-ui`}
+            Use ‘<Code>djy7snj</Code>’ for the{" "}
+            <Link href="https://helpx.adobe.com/uk/fonts/using/embed-codes.html">
+              Adobe Fonts embed code
             </Link>
-            {` generic font family.`}
+            . This code is valid for all of Barnardo’s products.
+          </Paragraph>
+          <Heading>Body copy</Heading>
+          <Paragraph>
+            You should use the{" "}
+            <Link href="https://www.w3.org/TR/css-fonts-4/#system-ui-def">
+              system-ui
+            </Link>{" "}
+            generic font family.
           </Paragraph>
           <Pagination
             previous={{