+++ src/pages/standards/voice-and-tone.js
@@ -16,7 +16,7 @@ import Topic from "../../components/Topic";
 export default () => (
   <Page>
     <Helmet>
-      <title>{`Voice and tone`}</title>
+      <title>Voice and tone</title>
     </Helmet>
     <Breadcrumbs
       items={[
@@ -31,41 +31,51 @@ export default () => (
       ]}
     />
     <Main>
-      <Topic>{`Writing`}</Topic>
-      <Title>{`Voice and tone`}</Title>
-      <Paragraph>{`When writing content be consistent, professional and effective to emotionally engage users in the
-        Barnardo’s brand.`}</Paragraph>
-      <Paragraph>{`To make content easy to find, use and understand you should:`}</Paragraph>
+      <Topic>Writing</Topic>
+      <Title>Voice and tone</Title>
+      <Paragraph>
+        When writing content be consistent, professional and effective to
+        emotionally engage users in the Barnardo’s brand.
+      </Paragraph>
+      <Paragraph>
+        To make content easy to find, use and understand you should:
+      </Paragraph>
       <BulletedList>
-        <ListItem>{`write in plain English`}</ListItem>
-        <ListItem>{`write in an active voice`}</ListItem>
-        <ListItem>{`choose words your audience would use`}</ListItem>
-        <ListItem>{`write short, clear sentences`}</ListItem>
-        <ListItem>{`use titles that are unique to the product`}</ListItem>
-        <ListItem>{`have the most important information first`}</ListItem>
-        <ListItem>{`check you can quickly tell what it’s about`}</ListItem>
-        <ListItem>{`check spelling, links and citations are correct`}</ListItem>
-        <ListItem>{`define acronyms and abbreviations the first time they’re used`}</ListItem>
-        <ListItem>{`explain technical and professional terms`}</ListItem>
-        <ListItem>{`optimise for search`}</ListItem>
+        <ListItem>write in plain English</ListItem>
+        <ListItem>write in an active voice</ListItem>
+        <ListItem>choose words your audience would use</ListItem>
+        <ListItem>write short, clear sentences</ListItem>
+        <ListItem>use titles that are unique to the product</ListItem>
+        <ListItem>have the most important information first</ListItem>
+        <ListItem>check you can quickly tell what it’s about</ListItem>
+        <ListItem>check spelling, links and citations are correct</ListItem>
+        <ListItem>
+          define acronyms and abbreviations the first time they’re used
+        </ListItem>
+        <ListItem>explain technical and professional terms</ListItem>
+        <ListItem>optimise for search</ListItem>
       </BulletedList>
-      <Paragraph>{`You should avoid:`}</Paragraph>
+      <Paragraph>You should avoid:</Paragraph>
       <BulletedList>
-        <ListItem>{`duplicating content`}</ListItem>
-        <ListItem>{`a passive voice`}</ListItem>
-        <ListItem>{`business or marketing speak`}</ListItem>
+        <ListItem>duplicating content</ListItem>
+        <ListItem>a passive voice</ListItem>
+        <ListItem>business or marketing speak</ListItem>
       </BulletedList>
-      <Paragraph>{`When writing for young people you should:`}</Paragraph>
+      <Paragraph>When writing for young people you should:</Paragraph>
       <BulletedList>
-        <ListItem>{`use an informal tone`}</ListItem>
-        <ListItem>{`be practical and actionable`}</ListItem>
-        <ListItem>{`lead with positive content`}</ListItem>
-        <ListItem>{`avoid being authoritarian, for example provide ‘tips’ not ‘rules’`}</ListItem>
+        <ListItem>use an informal tone</ListItem>
+        <ListItem>be practical and actionable</ListItem>
+        <ListItem>lead with positive content</ListItem>
+        <ListItem>
+          avoid being authoritarian, for example provide ‘tips’ not ‘rules’
+        </ListItem>
       </BulletedList>
       <Cite>
-        {`View `}
-        <Link href="https://github.com/barnardos/design-system/issues/510">{`voice and tone research`}</Link>
-        {` on GitHub.`}
+        View{" "}
+        <Link href="https://github.com/barnardos/design-system/issues/510">
+          voice and tone research
+        </Link>{" "}
+        on GitHub.
       </Cite>
       <Pagination
         next={{